// src/types/room-detail.ts
export interface RoomDetail {
  roomId: number
  roomDetailId?: number
  ettlDeviceId?: number
  deviceid?: number
  campusId?: number
  buildingId: number
  buildingName?: string
  buildingCode?: string
  blockId: number
  blockName?: string
  blockCode?: string
  floorId: number
  floorName?: string
  floorNo?: number
  roomName: string
  roomCode: string
  roomStrength?: number
  serialNumber?: string
  environment?: string
  isActive: boolean
  reason?: string
}